refactor(core): reuse collectUpstreamTaskIds for dependency output lookup
check-task-files already walks a task's dependencies with the same
direct-or-transitive switch. Lift it into task-graph-utils and use it
for getDependenciesWithOutputsToHash instead of a second traversal.
